requirements.txt
python-telegram-bot
python-telegram-bot[job-queue]
streamlit
pandas 
requests
nltk
wordcloud
cohere